# Implement SSE and AVX2 encoding