--- a/src/lib.rs
+++ b/src/lib.rs
@@ -49,6 +49,7 @@
 //!
 //! The `_slice` flavors of encode or decode will panic if the provided output slice is too small,
 
+#![cfg_attr(feature = "cargo-clippy", allow(cast_lossless))]
 #![deny(
     missing_docs,
     trivial_casts,
@@ -95,16 +96,16 @@ pub enum CharacterSet {
 }
 
 impl CharacterSet {
-    fn encode_table(&self) -> &'static [u8; 64] {
-        match *self {
+    fn encode_table(self) -> &'static [u8; 64] {
+        match self {
             CharacterSet::Standard => tables::STANDARD_ENCODE,
             CharacterSet::UrlSafe => tables::URL_SAFE_ENCODE,
             CharacterSet::Crypt => tables::CRYPT_ENCODE,
         }
     }
 
-    fn decode_table(&self) -> &'static [u8; 256] {
-        match *self {
+    fn decode_table(self) -> &'static [u8; 256] {
+        match self {
             CharacterSet::Standard => tables::STANDARD_DECODE,
             CharacterSet::UrlSafe => tables::URL_SAFE_DECODE,
             CharacterSet::Crypt => tables::CRYPT_DECODE,