@ggriffiniii 2018-11-20
Sorry, I realized when I rebased this pull request I reintroduced a previously fixed bug. I'll open a new pull request that fixes it. Sorry for the noise.